requirements.txt
cloudscraper
beautifulsoup4
python-telegram-bot==13.15